Prints "Hello, Go!" in the stdout
A Go app lives in a package called "main" and starts with the "main"
function, which is automatically invoked by the runtime.
The function declaration syntax is relatively closed to C or Java.
Here, the particular main function takes no arguments and returns nothing.

In this basic example app, we rely on the "fmt" package which we
import on line 3 from the standart library, to print the string "Hello, Go!"
in the standart output. Once again, the syntax for invoking a function is
closed to C or Java.

Try the command "go run main.go" to run the program from your command
line.