Starts a simple HTTP server which handles /hello
The application now uses the "net/http" package from the standart library
to create a simple HTTP server. Note that once "net/http" is imported,
its public API is made accessible through the "http" variable.

Basically we are doing two things here:

Firstly between lines 9 and 11, we register a handler for the "/hello" route.
The handler is a function which we can either be defined separatly or as a
lexical closure.
In Go, functions are first class values and can be passed as parameters in
other functions which is why you can do this.
However, since Go is strongly typed, this particular function must have a
specific signature in order for the program to compile.
The signature defines two arguments: a http.ResponseWriter which we use
to construct the HTTP response, and a *http.Request which is a pointer
to a variable of type http.Request. If you are not familiar with pointers,
you can see them as Java references.
Note that we use w.Write which takes a slice of bytes, to create the
response body. By default, the server will respond with the HTTP 200
status code.

Secondly on line 13, we invoke http.ListenAndServe and provide a TCP port
to wire up.
At this point, the program will block and start listening to incoming HTTP
requests. With the second argument we can provide a custom HTTP mutex to
handle concurrent requests. In this case, we want to use the default with
is contained in the http package. To do that we simply provide "nil" as a
second argument.

Note that this server really is a production ready server. Where with
other technologies you would need an application server or a separate HTTP
server to serve your code, with Go you don't really that.

Until now you used "go run main.go" to compile and run the code. Something
you will need to do, is to build your code so you can ship it to
production (or whatever environment).
To do that, simply run "go build", the command will generate an statically
linked binary file which contains the application and the runtime.

This executable file is self contain, which means it has no dependencies.
If you ran "go build" on a Windows 64 bits system, you can run it on
another Windows machine which does not have Go installed and yet it will
run.
Since it does not depend on any virtual machine or pre-configured
environment, your Windows binary will not run on a Mac or a Linux based
operating system. That being said, Go provides a great support for
cross-compilation.