Creates a separate package for our HTTP handlers
In most cases, a Go app is splited in multiple packages.
Though in this case, the code base is still quite small, a new "api"
package is created for our HTTP handlers in order to illustrate the
principle.

Basically in our workspace, each directory represents a package that
contains one or many files.
The "api" package sources are stored in the "api" sub-directory, and
each source file have to describe "package api".

Note that in "api/repos.go", the handler for "/hello" was moved and
given the name "Hello". The capitalized name means that the function is
exported from the package, thus package users will be able to use the
"Hello" function. If we named the function "hello", then it would be
unexported. This rule applies to everything in a package (functions,
variables, types, etc...).

In "main.go" on line 7, the "api" package is exported given its path
relative to the $GOPATH/src folder, and "api.Hello" is now passed as
the second argument of "http.HandleFunc".