Uses dependencies injection to decouple our code
The logic for fetching repositories from Github and cloning remote
repositories will be exported in a dedicated package named "github".

If we were coding this app in Java, we would propably create a
"ReposManager" class, this class would be used by the HTTP handler which
would probably become a method of a controller.

Go does not provide classes, but we can use structures to reproduce the
same mechanisms. In "github/github.go" on line 8, we create a new
structure.
Technically, what we really do here is, we create a new "ReposManager" type
which is based on type struct{}, which is an empty data structure.
It worth noting that Go does not have type inheritance, thus: ReposManager !== struct{}

Between lines 10 and 16, we create two functions that are scoped in our
new "ReposManager" type, which technically makes them methods. For now
we leave those methods empty and we will implement them later.

According to you, what will happen if we don't capitalize names for
those two methods?

In "api/repos.go", we changed the code and created a new structure named
"ReposHandlers". The HTTP handler is now a method of this new type.
As you can see on line 15, "ReposHandlers" has a dependency of something
of type "listCloner". This type is defined between lines 9 and 12 as an
interface.
Interfaces are a key feature of Go, any type which implements all the
functions listed in an interface automatically implements this interface
(without the need to explicitly "implements" it) and can be used in code
bases that depend on this interface.

This is good because it makes it easy to merge two code bases that have
no types in common.

Now in "main.go" on lines 12 and 13, we need to construct our dependency
graph and adapt the second argument of "http.HandleFunc".