Implements "List" and "Clone" for "github.ReposManager"
The implementation for "List" makes use of the default HTTP client from
the standart library to perform a get request to get the repositories
of a given Gitbub user. The list is represented as an array of JSON
objects. The default JSON decoder from the standart library is used to
extract the JSON stream into a Go value, in this case: a slice of "Repo".

Note the strings that annotate the fields for the "Repo" structure on
lines 11 and 12?
This specific Go feature is called "structure field's tags". Field's tags can
be used in many ways through the "reflect" package.
The "encoding/json" package makes use of field's tags to bind JSON fields
name to Go structure fields.
Another good example is the Gorm ORM which attach database table fields
to Go structures using field's tags.

The implementation for "Clone" makes use of the "os/exec" package from
the standart library to perform a "git" command on the operating system.

We changed both signatures in order to be able to return an "error". This
makes sense since both functions perform I/O operations that may
potentially fail.
Go does not provide exceptions, generally speaking, it considers errors as
non-exceptionnal and treat them as regular values.
Instead, Go offers the ability for a function to return multiple values.
Typically this is the way errors are managed, a function that could
fail returns two values: the first one the nominal value, the second one
is an error. If the error is not nil, it means that something went wrong
and we have to deal with it.

This can lead to some code duplications, but this is a
trade-off that also brings value from a readability perpective.

Obviously as a consequence, the code in "api/repos.go" is adapted. For
now, errors are simply printed out in the response body.