Clones all repositories concurrently
In Go, concurrent jobs are executed in goroutines, they can be seen as
light threads that are automatically managed by the go runtime.

By default, invoking a function will block the execution flow.
Using the "go" keyword as we do in "api/repos.go" on line 29, the provided
function will execute in a goroutine, and thus will not block the
execution flow anymore.

We start as many concurrent goroutines as there are repositories, which
brings the need for synchronizing them.
In fact, we don't want the instruction on line 43 to be executed before
all goroutines are done.

Goroutines can exchange messages through channels. Channels are generic
built-in types. On line 27, we create a new channel of booleans with of
buffer size equals to the number of repositories.

On line 40, we read the next value from the channel, this operation
will block the execution flow until a value is passed in this channel.
We pass new values in the channel each time a repository is done cloning
on line 36.

In consequence, the program will not exit the loop defined on line 39
until the last repository is done cloning.